Extract the boiler plate code into a factory
Manage the query history store lifecycle in the factory.
