+++ presto-main/src/main/java/com/facebook/presto/execution/QueryTracker.java
@@ -14,11 +14,10 @@
 package com.facebook.presto.execution;
 
 import com.facebook.presto.Session;
-import com.facebook.presto.server.extension.ExtensionFactory;
 import com.facebook.presto.server.extension.query.history.QueryHistoryStore;
+import com.facebook.presto.server.extension.query.history.QueryHistoryStoreFactory;
 import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.QueryId;
-import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
 import io.airlift.log.Logger;
 import io.airlift.units.Duration;
@@ -27,14 +26,9 @@ import org.joda.time.DateTime;
 import javax.annotation.concurrent.GuardedBy;
 import javax.annotation.concurrent.ThreadSafe;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.Collection;
 import java.util.NoSuchElementException;
 import java.util.Optional;
-import java.util.Properties;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
@@ -82,41 +76,7 @@ public class QueryTracker<T extends QueryExecution>
         this.clientTimeout = queryManagerConfig.getClientTimeout();
 
         this.queryManagementExecutor = requireNonNull(queryManagementExecutor, "queryManagementExecutor is null");
-        queryHistoryStore = loadQueryHistoryExtension();
-    }
-
-    private Optional<? extends QueryHistoryStore> loadQueryHistoryExtension()
-    {
-        Properties extensionProps;
-        try {
-            extensionProps = getExtensionConf();
-        }
-        catch (IOException e) {
-            log.warn("Failed to load query extension config from " + QUERY_HISTORY_EXTENSION_CONFIG_FILE);
-            return Optional.empty();
-        }
-        if (extensionProps == null) {
-            return Optional.empty();
-        }
-        // The implementation class is defined as a property `com.facebook.presto.server.extension.query.history.QueryHistoryStore.impl`.
-        String extensionImplClass = extensionProps.getProperty(QueryHistoryStore.class.getName() + ".impl");
-        if (Strings.isNullOrEmpty(extensionImplClass)) {
-            return Optional.empty();
-        }
-        return ExtensionFactory.INSTANCE.createExtension(extensionImplClass, extensionProps, QueryHistoryStore.class);
-    }
-
-    private static Properties getExtensionConf() throws IOException
-    {
-        File extensionPropsFile = new File(QUERY_HISTORY_EXTENSION_CONFIG_FILE);
-        if (extensionPropsFile.exists()) {
-            Properties config = new Properties();
-            try (InputStream configResource = new FileInputStream(extensionPropsFile)) {
-                config.load(configResource);
-            }
-            return config;
-        }
-        return null;
+        queryHistoryStore = QueryHistoryStoreFactory.getQueryHistoryStore();
     }
 
     public synchronized void start()
